properly sized image from cloudinary and styled homepage boxes

+++ db/migrate/20201205213243_change_price_in_items.rb
@@ -1,5 +1,5 @@
 class ChangePriceInItems < ActiveRecord::Migration[6.0]
   def change
-    change_column :items, :price, :float
+    change_column :items, :price, :integer
   end
 end
